- combine the download facility download viewer page
